refactor: flatten tools directory structure by removing unnecessary schemas subdirectory
🏗️ Directory Structure Flattening:
- Move all schema files from tools/schemas/ directly to tools/
- Remove unnecessary tools/schemas/ subdirectory and its __init__.py
- Update import in tool_registry.py from .tools.schemas to .tools
- Merge schemas/__init__.py functionality into tools/__init__.py

📁 File Changes:
- ray_mcp/tools/schemas/cloud_tools.py → ray_mcp/tools/cloud_tools.py
- ray_mcp/tools/schemas/cluster_tools.py → ray_mcp/tools/cluster_tools.py
- ray_mcp/tools/schemas/job_tools.py → ray_mcp/tools/job_tools.py
- ray_mcp/tools/schemas/log_tools.py → ray_mcp/tools/log_tools.py
- ray_mcp/tools/schemas/schema_utils.py → ray_mcp/tools/schema_utils.py
- Remove ray_mcp/tools/schemas/__init__.py (merged into tools/__init__.py)

✅ Benefits:
- Simpler directory structure (eliminates unnecessary nesting)
- Cleaner imports (from .tools instead of .tools.schemas)
- Easier navigation (all tool files in one directory)
- Reduced complexity (fewer __init__.py files to maintain)
- All 245 tests pass - zero functionality impact

💡 Rationale:
- tools/schemas/ was the only subdirectory under tools/
- Closely related files don't need additional organizational layer
- Follows principle of least structure - only add directories when they provide clear value
