+++ ray_mcp/tools/__init__.py
@@ -1,6 +1,6 @@
-"""Ray MCP tools package with modular schemas."""
+"""Ray MCP tools package with tool schemas organized by functionality."""
 
-from .schemas import cloud_tools, cluster_tools, job_tools, log_tools
+from . import cloud_tools, cluster_tools, job_tools, log_tools
 
 __all__ = [
     "cluster_tools",
