--- ray_mcp/tools/schemas/__init__.py
@@ -1,10 +1,0 @@
-"""Tool schemas for Ray MCP server organized by functionality."""
-
-from . import cloud_tools, cluster_tools, job_tools, log_tools
-
-__all__ = [
-    "cluster_tools",
-    "job_tools",
-    "cloud_tools",
-    "log_tools",
-]
